misc:stml4xx: Extend SensorHub WAKE_SENSORS array to 24 bits -- L4.
All 16 sensor wake flags have been used. We need to expand the number of
available bits to implement new gestures.

Reviewed-on: http://gerrit.mot.com/729511
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ drivers/misc/moto-sensorhub/motosh_core.c
@@ -72,7 +72,7 @@ unsigned short motosh_g_step_counter_delay;
 unsigned short motosh_g_ir_gesture_delay;
 unsigned short motosh_g_ir_raw_delay;
 unsigned long motosh_g_nonwake_sensor_state;
-unsigned short motosh_g_wake_sensor_state;
+unsigned long motosh_g_wake_sensor_state;
 unsigned short motosh_g_algo_state;
 unsigned char motosh_g_motion_dur;
 unsigned char motosh_g_zmotion_dur;
